+++ src/v2_0_1/datatypes/charging_profile_criterion_type.rs
@@ -13,10 +13,10 @@ pub struct ChargingProfileCriterionType {
     pub charging_profile_purpose: Option<ChargingProfilePurposeEnumType>,
     /// Optional. Value determining level in hierarchy stack of profiles. Higher values have precedence over lower values. Lowest level is 0
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub stack_level: Option<i64>,
+    pub stack_level: Option<i32>,
     /// Optional. List of all the chargingProfileIds requested. Any ChargingProfile that matches one of these profiles will be reported. If omitted, the Charging Station SHALL not filter on chargingProfileId. This field SHALL NOT contain more ids than set in ChargingProfileEntries.maxLimit
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub charging_profile_id: Option<Vec<i64>>,
+    pub charging_profile_id: Option<Vec<i32>>,
     /// Optional. For which charging limit sources, charging profiles SHALL be reported. If omitted, the Charging Station SHALL not filter on chargingLimitSource.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub charging_limit_source: Option<Vec<ChargingLimitSourceEnumType>>,
