+++ src/v2_0_1/datatypes/clear_charging_profile_type.rs
@@ -7,11 +7,11 @@ use crate::v2_0_1::enumerations::charging_profile_purpose_enum_type::ChargingPro
 pub struct ClearChargingProfileType {
     /// Optional. Specifies the id of the EVSE for which to clear charging profiles. An evseId of zero (0) specifies the charging profile for the overall Charging Station. Absence of this parameter means the clearing applies to all charging profiles that match the other criteria in the request.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub evse_id: Option<i64>,
+    pub evse_id: Option<i32>,
     /// Optional. Specifies to purpose of the charging profiles that will be cleared, if they meet the other criteria in the request.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub charging_profile_purpose: Option<ChargingProfilePurposeEnumType>,
     /// Optional. Specifies the stackLevel for which charging profiles will be cleared, if they meet the other criteria in the request.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub stack_level: Option<i64>,
+    pub stack_level: Option<i32>,
 }
