+++ src/v2_0_1/datatypes/composite_schedule_type.rs
@@ -8,8 +8,8 @@ use crate::v2_0_1::enumerations::charging_rate_unit_enum_type::ChargingRateUnitE
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
 #[serde(rename_all = "camelCase")]
 pub struct CompositeScheduleType {
-    pub evse_id: i64,
-    pub duration: i64,
+    pub evse_id: i32,
+    pub duration: i32,
     pub schedule_start: DateTime<Utc>,
     pub charging_rate_unit: ChargingRateUnitEnumType,
     pub charging_schedule_period: Vec<ChargingSchedulePeriodType>,
