+++ src/v2_0_1/datatypes/dc_charging_parameters_type.rs
@@ -4,17 +4,17 @@ use validator::Validate;
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Validate, Default)]
 #[serde(rename_all = "camelCase")]
 pub struct DCChargingParametersType {
-    pub ev_max_current: i64,
-    pub ev_max_voltage: i64,
+    pub ev_max_current: i32,
+    pub ev_max_voltage: i32,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub energy_amount: Option<i64>,
+    pub energy_amount: Option<i32>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub ev_max_power: Option<i64>,
+    pub ev_max_power: Option<i32>,
     #[serde(skip_serializing_if = "Option::is_none")]
     #[validate(range(min = 0, max = 100))]
     pub state_of_charge: Option<u8>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub ev_energy_capacity: Option<i64>,
+    pub ev_energy_capacity: Option<i32>,
     #[serde(rename = "fullSoC")]
     #[serde(skip_serializing_if = "Option::is_none")]
     #[validate(range(min = 0, max = 100))]
