+++ src/v2_0_1/datatypes/evse_type.rs
@@ -3,7 +3,7 @@
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
 #[serde(rename_all = "camelCase")]
 pub struct EVSEType {
-    pub id: i64,
+    pub id: i32,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub connector_id: Option<i64>,
+    pub connector_id: Option<i32>,
 }
