+++ src/v2_0_1/datatypes/variable_monitoring_type.rs
@@ -5,9 +5,9 @@ use crate::v2_0_1::enumerations::monitor_enum_type::MonitorEnumType;
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Default)]
 #[serde(rename_all = "camelCase")]
 pub struct VariableMonitoringType {
-    pub id: i64,
+    pub id: i32,
     pub transaction: bool,
-    pub value: f64,
+    pub value: f32,
     #[serde(rename = "type")]
     pub kind: MonitorEnumType,
     pub severity: u8,
