+++ src/v2_0_1/messages/clear_variable_monitoring.rs
@@ -6,7 +6,7 @@ use crate::v2_0_1::datatypes::clear_monitoring_result_type::ClearMonitoringResul
 #[serde(rename_all = "camelCase")]
 pub struct ClearVariableMonitoringRequest {
     /// List of the monitors to be cleared, identified by there Id.
-    pub id: Vec<i64>,
+    pub id: Vec<i32>,
 }
 
 /// ClearVariableMonitoringResponse, sent by the Charging Station to the CSMS.
