+++ src/v2_0_1/messages/publish_firmware_status_notification.rs
@@ -8,7 +8,7 @@ pub struct PublishFirmwareStatusNotificationRequest {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub location: Option<Vec<String>>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub request_id: Option<i64>,
+    pub request_id: Option<i32>,
 }
 
 /// This contains the field definition of the PublishFirmwareStatusNotificationResponse PDU sent by the CSMS to the Charging station in response to a PublishFirmwareStatusNotificationRequest.
